peer: Add inv type summary to debug message.
This modifies the debug message summary for inv messages that contain a
list with multiple items to include a count of the blocks and
transactions as opposed to solely the size of the list.  This is useful
information for the new sync model to help confirm there are no blocks
in the inv messages.