mining: No error log on expected head reorg errors.
This changes the log for head reorg failures to a debug log instead of
an error log since it is actually expected to fail reorgs from time to
time due to the fact mining is a race to find blocks.

It also cleans up the code slightly while here and adds an additional
check to avoid additional work in the case reorgs to other tips fail and
the current tip then becomes the best candidate among the eligible
parents.