rpcserver: Fix count tspend votes in mined block
This fixes a bug where requesting the treasury spend vote counts on the
exact block where it was mined via the gettreasuryspendvotes rpc call
would fail with "tspend outside window".

Explicitly passing the block where the tspend was mined when that block
was the last possible block that particular tspend could be included in
the blockchain would fail because the votes on that block themselves
aren't counted towards determining tspend elligibility: votes are
tallied only up to its parent.

In essence, the test determining the cutoff height had an off-by-one by
using > instead of >= when a block was explicitly specified in the
request.