blockchain: Add utxoSetState to the database.
This adds a utxoSetState type that is tracked in the database. The utxo
set state contains information regarding the current state of the utxo
set.  In particular, it tracks the block height and block hash of the
last completed flush.

The utxo set state is tracked in the database since at any given time,
the utxo cache may not be consistent with the utxo set in the database.
This is due to the fact that the utxo cache only flushes changes to the
database periodically.  Therefore, during initialization, the utxo set
state is used to identify the last flushed state of the utxo set and it
can be caught up to the current best state of the main chain.

This additionally adds full test coverage to the new serialization and
deserialization functions.