wire: Deprecate reject message.
This deprecates the reject wire protocol message by bumping the protocol
version to 9 and making the reject message illegal under the new
protocol version.  It also bumps the default user agent to
dcrwire:1.0.0.

Note that the message is not removed yet because the code still has to
be able to negotiate to older protocols where it is still supported
until some future software version makes the entire network require the
new protocol version is deployed at which time the entire message can be
removed.

This is part of deprecating the reject wire protocol message towards its
eventual removal in a future version of the software.