server: Respond to getheaders when same chain tip.
The handler for getheaders currently ignores the request when the chain
is not yet believed to be current.  This is generally desirable and
correct behavior, since it might otherwise lead peers to incorrect
conclusions about the state of the peer.

However, on private networks, such as simnet, it is not at all uncommon
for every node in the network to no longer be current if a block hasn't
been mined in a long time or when all nodes are stopped and restarted.

In order to better handle these types of edge conditions, this modifies
the server to respond to getheaders when the local chain tip is exactly
the same as the requested locator even when it is not marked current
yet.

This results in more robust handling for private networks while still
providing the normal desirable behavior prior to being current.