blockchain: Deep copy view entry script from tx.
This updates utxo viewpoints to deep copy the script when getting it
from a msg tx.  This is required since the tx out script is a subslice
of the overall contiguous buffer that the msg tx houses for all scripts
within the tx.  It is deep copied here since this entry may be added to
the utxo cache, and we don't want the utxo cache holding the entry to
prevent all of the other tx scripts from getting garbage collected.