blockchain: Add UtxoCacher interface.
This adds a UtxoCacher interface so that alternative utxo cache
implementations can be provided.  In particular, this will be used to
provide a mock implementation for testing in order to more easily
simulate various scenarios.

In addition to introducing the UtxoCacher interface, this updates
BlockChain and UtxoViewpoint to use the interface rather than the
concrete type.