blockchain: Add UtxoCache Initialize tests.
This adds tests for the UtxoCache Initialize method.  These tests
include recovery scenarios that are particularly hard to simulate to
ensure that those code paths are not broken in the future.

In order to test recovery scenarios, this introduces a test utxo cache
that allows for toggling flushing on and off.

Additionally, this adds an ExpectUtxoSetState method to chaingenHarness
that allows for easily validating the last flushed block for the utxo
set.