stdaddr: Add infrastructure for v0 decoding.
This adds the overall infrastracture for decoding version 0 addresses
along with associated testing infrastructure.

Note that, in order to help ease the review process, this only adds the
relevant infrastructure and does not yet support any version 0 address
types.  Each supported version 0 address type will be added in future
commits.

This is part of a series of commits to fully implement the stdaddr
package.