txscript: Add script bldr unchecked op add.
This adds a new method to ScriptBuilder named AddOpsUnchecked that is
similar to AddDataUnchecked in that it allows the limits that are
ordinarily impose to be bypassed for regression testing purposes where
sizes are intentionally made larger than allowed.

It also updates the test code to make use of the method instead of
reaching into the internals to manually concatenate the opcodes.