blockchain: Allow alternate tips for current check.
This updates the check which determines whether the chain believes it is
current to also detect when the current best chain tip is an alternative
tip at the same height as the best header.

This is being done because the best header is determined based on when
the block data was received during steady state operation to prevent
miners being able to gain an unfair advantage, however, that order is
intentionally not stored to the database since it only applies to steady
state.  The result is that restarting the software when there are
multiple chain tips might result in one of the other variants becoming
the best header as compared to when it was shutdown and thus it will not
match the current best block.