stdaddr: Accept vote and revoke limits separately.
This modifies the RewardCommitmentScript method of the StakeAddress
interface to accept the vote and revoke fee limits separately instead of
expecting the caller to have already encoded it properly.

This simplifies the creation of the reward commitment script since there
is not currently any other method available anywhere to independently
encode the limits properly and it is not a particularly trivial
encoding.

Finally, the tests and example are updated accordingly.