indexers: Convert to use new stdaddr package.
This converts the blockchain/indexers package to use the stdaddr package
instead of dcrutil.Address.

All callers that use the package in the repository are updated
accordingly to maintain code that continues to build and work properly
as well as pass all tests.  In order to achieve this, it introduces
temporary address conversion code that will be removed in future
commits.

This is part of a series of commits to convert all packages in the
repository to make use of the new stdaddr package.