addrmgr: Add network address type filter.
This commit introduces a network address type filter that allows a
caller to indicate the types of addresses that can be returned from
the address manager.  This allows finer control of what types of
addresses may be broadcast to a particular peer.  The filters are
decided by the protocol version of the peer in a way that allows the
address manager's internal implementation to remain agnostic of the
protocol versioning logic.