blockchain: Decouple stxo and utxo migrations.
This decouples the overall block database version and migrations from
the spend journal and utxo set versions and migrations.

This provides a couple of benefits.  Firstly, the migration to version 3 of the
spend journal and utxo set currently require the block index to be loaded.  In
the future this could be handled by reindexing, once that is introduced, but
until then this decoupling allows for new block database versions and migrations
to be run prior to the block index being loaded.

Secondly, this is a step toward removing the dependency that the utxo set and
spend journal have on the block database and index.

This is part of moving the UTXO set and state to a separate database.