multi: Migrate to UTXO database.
This migrates the UTXO set and state from the block database to the new
UTXO database.  The benefit of doing this is that it will allow the
reading and writing to the UTXO set in the database to be optimized
specifically for use with the UTXO set.

An overview of the changes is as follows:

- Introduce upgrade code to migrate UTXO data from the block database to
  the new UTXO database
- Update all references to use the UTXO database rather than the block
  database where appropriate