blockchain: Trsy always inactive for genesis blk.
This modifies the IsTreasuryAgendaActive method to return false when a
zero hash is provided since that indicates the next block is the genesis
block and the treasury agenda is always inactive for the genesis block.

This has the effect of fixing a few RPC handlers, such as getblock, when
working with the genesis block.