contrib: Update OpenBSD rc script for 6.9 features.
OpenBSD 6.9 includes a new daemon_logger variable which may be set to
automatically set the syslog facility that the process will write to.
This lets us simplify our example service by avoiding the need to
redefine the rc_start function.

While here, update some of the default flags to more closely match
what I now recommend.  /var is too small for block data on many
installations, so it is better to write all of the application data to
the /home partition instead.  This also fixes the comment for how to
enable the service, since just defining the flags is not sufficient
for this.