multi: Flush block DB before UTXO DB.
This adds a flushBlockDB function to the UTXO cache.  The flushBlockDB
function is used to flush the block database to disk prior to flushing
the UTXO cache to the UTXO database.

This ensures that the block database is always at least as far along as
the UTXO database which keeps the UTXO database in a recoverable state
in the event of an unclean shutdown.