blockchain: Force flush in separateUtxoDatabase.
This modifies the separateUtxoDatabase upgrade to force the UTXO
database to flush to disk prior to removing the UTXO set and state from
the block database.

This prevents a scenario where the UTXO set is removed from the block
database and the block database is flushed to disk, but an unclean
shutdown occurs before the UTXO database flushes to disk, leaving the
UTXO database in an unrecoverable state.