blockchain: Add UtxoBackend interface.
This adds a UtxoBackend interface and LevelDbUtxoBackend type, which
implements that interface.  This UtxoBackend interface will allow
different backend implementations to be used by the UtxoCache rather
than interacting directly with the database layer as it does currently.

In subsequent commits, interface methods and the corresponding
implementation will be added to UtxoBackend one at a time in order to
ease the review process as things move around.