stdscript: Introduce pkg infra for std scripts.
Similar to the code for handling standard addresses prior to the recent
updates which implemented the stdaddr package, the current code for
handling standard scripts implemented in txscript was written many years
ago prior a wide variety of changes and several new features added by
Decred.  As a result, it entirely lacks support for some features and
supports others in a roundabout and non-intuitive way.

Specifically, it does not support or provide a clean path to enable
support for different script versions.

Finally, the code being implemented in txscript (which was necessary
back when the original code was implemented, but no longer is due to
changes since) has led to confusion regarding what is considered
standard and what is considered consensus which has tripped up several
contributors over the years.

This is part of a series of commits that aims to resolve all of the
aforementioned issues by introducing a new package named stdscript which
reworks the way standard scripts are handled.

For the time being, the package is introduced into the internal staging
area for initial review.

The following provides an overview of some of the key features of the
new design:

- Supports versioned scripts
- Provides the ability to efficiently extract version-specific relevant
  data, such as public keys, public keys hashes, and script hashes from
  standard scripts
- Allows callers to choose between using version-specific or dynamic
  type determination methods
- Implements several new types, as compared to the existing code, which
  leads to an easier to use API
- Clearly denotes that these scripts are a standardized construction the
  must not be used directly in consensus code
- Provides additional convenience methods for working with typical
  version 0 atomic swap and multisignature scripts

In order to help ease the review process, this commit only contains the
overall generic infrastructure without adding support for any specific
script types.  Each supported version 0 script type will be added in
future commits.