stdscript: Add v0 p2pk-schnorr-secp256k1 support.
This adds support for detecting version 0
pay-to-pubkey-schnorr-secp256k1 scripts as standard and extracting the
associated public key.  Note that consensus only allows compressed
public keys for this signature type so only compressed public keys are
detected and extracted.

Full test coverage is included.

This is part of a series of commits to fully implement the stdscript
package.