stdscript: Add v0 ecdsa multisig support.
This adds support for detecting version 0 ecdsa multisignature scripts
as standard and extracting the associated details, including the number
of required signatures, number of public keys, and the public keys
themselves.

It should be noted that this code includes a change in the semantics as
compared to the existing code.  In particular:

- The new code only multisignature scripts with compressed public keys
  as standard versus the existing code that allows both compressed and
  uncompressed public keys
- The new code requires at least one signature to be required versus the
  existing code that only requires any small integer (which includes 0)

These ensure the scripts take up considerably less space on chain and
there really isn't a reasonable justification for not using compressed
public keys or not requiring any signatures since that could be done
with a smaller script if truly desired.

In practice, all known software already conforms to these semnatics and
has since launch.

This, of course, does not affect consensus in any way and it still
technically supports both uncompressed and compressed public keys as
well as zero required signatures for the associated multisig opcode.

Full test coverage is included.

This is part of a series of commits to fully implement the stdscript
package.