stdscript: Add v0 atomic swap redeem support.
This adds support for identifying if a redeem script is a version 0
atomic swap redeem script for an original pay-to-script-hash script as a
convenience method for consumers of the package.

Note that this is different than most other methods in the package in
that it applies to the redeem script as opposed to the public key script
and consequently there is no new script type introduced nor does the
overall script determination detect it.  This is the case because a
redeem script is the final push of a signature script that is itself a
redemption of a pay-to-script-hash script.

Full test coverage is included.

This is part of a series of commits to fully implement the stdscript
package.