blockchain: Use scripts in tickets address query.
This modifies the TicketsWithAddress query method in blockchain to
accept the more specific stdaddr.StakeAddress since only stake-capable
addresses can be used in tickets and modifies the logic to compare the
expected voting rights script and associated script version directly
instead of extracting and comparing encoded addresses.  It also updates
the callers and plumbing in rpcserver accordingly as well as adds a test
for the new non-stake address error in the RPC handler.

Not only is this approach more efficient, it also removes the reliance
on address extraction which in turn removes the need for the flag which
specifies whether or not the treasury is active.