stake: Do not use standardness code in consensus.
This modifies the treasury code in stake to avoid using standardness
code since it is consensus level code which mean it must not change even
though the policy regarding standardness can change.

The consensus-specific methods are all already available in the stake
package, so the code is merely updated to use them instead.

An overview of the changes follows:

- Introduce an exported constant for the max data carrier size allowed
  in stake transactions set to match the current standardness value
  since that is now part of consensus for them
- Modify the calls to txscript.GetScriptClass used for determining null
  data scripts to use consensus-specific IsNullDataScript method
- Use the consensus-specific methods for the following:
  - IsStakeChangeScript
  - isPubKeyHashScript
  - isScriptHashScript