blockchain: Remove unneeded OP_TADD maturity check.
This removes the unnecessary OP_TADD coinbase maturity check because
OP_TADD is not a spendable utxo and therefore it can never have a
maturity applied to it.  Recall that treasurybase and treasury add stake
transactions are the only place an OP_TADD is valid and in both cases
they result in increasing the balance of the treasury account and are
never directly spendable.