stake: Comment on max SSGen outputs with treasury.
This adds a comment to MaxOutputsPerSSGen indicating that if the
treasury agenda is active, an OP_RETURN TV (treasury vote) output is
additionally included in the last position when voting on a treasury
spend.

The MaxOutputsPerSSGen value does NOT account for this output, and
should technically be increased by 1, but since the treasury agenda is
already active this cannot be updated without a consensus vote.