multi: Move UTXO database.
This moves the UTXO database from a "metadata" subdirectory to the
parent directory.  The UTXO database existed in a "metadata"
subdirectory during the transition phase of moving the UTXO set to a
separate database since it was using the database package, but the
subdirectory is no longer necessary now that the UTXO database uses
leveldb directly.

Additionally, it bumps the overall block database version to 11 in order
to prevent downgrades.