blockchain: Add IsTreasuryEnabled to AgendaFlags.
This adds an IsTreasuryEnabled method to the AgendaFlags type that
returns whether the flags indicate that the treasury agenda is enabled.
This is useful since checking the flags directly is pretty verbose,
especially when using the type outside of the blockchain package.

This also updates existing references to the AgendaFlags type to use
the IsTreasuryEnabled function.