mempool: Add maybeAcceptTransaction AgendaFlags.
This updates the mempool maybeAcceptTransaction method to accept agenda
flags rather than just a boolean indicating if the treasury is active.

This allows for accommodating future agendas without having to add
additional boolean fields for each agenda.

This is useful in particular since maybeAcceptTransaction makes use of
the blockchain.CheckTransaction function to perform validation checks on
transactions, which includes checks which depend on whether or not an
agenda is active.