multi: Add auto revocations to agenda flags.
This adds AFAutoRevocationsEnabled to blockchain.AgendaFlags, which may
be set to indicate that the automatic ticket revocations agenda should
be considered as active when checking a transaction so that any
additional checks which depend on the agenda being active are applied.

Additionally, this updates all references of blockchain.AgendaFlags to
set AFAutoRevocationsEnabled if the automatic ticket revocations agenda
is active.